Toevoegen van Default mapping
Ik heb nu de default mappings toegevoegd met bijhorende tickrepository
statements
nu alleen nog powerpoint afmaken en klaar is kees
