+++ Supportcenter/DAL/packages.config
@@ -3,5 +3,5 @@
   <package id="EntityFramework" version="6.1.3" targetFramework="net45" />
   <package id="FluentNHibernate" version="2.0.3.0" targetFramework="net45" />
   <package id="Iesi.Collections" version="4.0.0.4000" targetFramework="net45" />
-  <package id="NHibernate" version="4.0.0.4000" targetFramework="net45" />
+  <package id="NHibernate" version="4.0.4.4000" targetFramework="net45" />
 </packages>
